classifier/img_classifier_model/requirements.txt
tensorflow
Flask
Pillow
scikit-learn
pandas
numpy
opencv-python
matplotlib